// packages/utils/src/date/type.ts
export interface DateInfo {
  year: number;
  month: number;
  day: number;
  dayStr: string;
  hour: number;
  hourStr: string;
  minute: number;
  minuteStr: string;
  second: number;
  secondStr: string;
  weekdayEn: string;
  weekdayCn: string;
  monthEn: string;
  monthCn: string;
  timestamp: number;
  dateStr: string;
}